(feat|COS-4200): Segment invoices table. User can see only one invoices item in side-bar and navigate between past and upcoming invoices via toggle button in table view
