--- packages/apps/frontera/src/routes/organizations/src/components/ContactOrgViewToggle/index.ts
@@ -1,1 +1,0 @@
-export { ContactOrgViewToggle } from './ContactOrgViewToggle.tsx';
