+++ packages/apps/frontera/src/routes/organizations/src/components/TableViewsToggleNavigation/index.ts
@@ -1,0 +1,1 @@
+export { TableViewsToggleNavigation } from './TableViewsToggleNavigation';
