+++ packages/apps/frontera/src/routes/src/components/RootSidenav/components/sections/index.tsx
@@ -1,0 +1,3 @@
+export * from './NavigationSections';
+export * from './LogoSection';
+export * from './UserActionSection';
