+++ packages/apps/frontera/src/routes/src/components/RootSidenav/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useNavigationManager';
+export * from './usePreferencesManager';
